Make mobs spawn offscreen
The distance they are spawned at from the screen borders is controlled
by the PAD constant.

Also had to rewrite the code responsible for setting the rotation
of the mob as path can no longer be used for this purpose.

Lastly, changed the random component of the rotation to PI/8,
instead of PI/4. Because the old range was too big when spawning so far
off the screen.